# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RouteWise API",
    "version": "0.1.0",
    "description": "Student mobility planning. Demo responses are fictional. Route IDs connect cards to route objects."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/places/nearby": {
      "post": {
        "summary": "Find nearby places, including parking",
        "tags": [
          "Places"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "location": {
                    "type": "object",
                    "properties": {
                      "latitude": {
                        "type": "number",
                        "minimum": -90,
                        "maximum": 90
                      },
                      "longitude": {
                        "type": "number",
                        "minimum": -180,
                        "maximum": 180
                      }
                    },
                    "required": [
                      "latitude",
                      "longitude"
                    ],
                    "additionalProperties": false
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "food",
                      "coffee",
                      "parking",
                      "gas",
                      "ev_charging",
                      "pharmacy",
                      "groceries",
                      "dessert"
                    ]
                  },
                  "radiusMeters": {
                    "default": 800,
                    "type": "integer",
                    "minimum": 100,
                    "maximum": 5000
                  },
                  "openNow": {
                    "default": false,
                    "type": "boolean"
                  }
                },
                "required": [
                  "location",
                  "category"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "security": [],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "dataMode": {
                      "type": "string",
                      "enum": [
                        "demo",
                        "live"
                      ]
                    },
                    "places": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "address": {
                            "type": "string"
                          },
                          "location": {
                            "type": "object",
                            "properties": {
                              "latitude": {
                                "type": "number",
                                "minimum": -90,
                                "maximum": 90
                              },
                              "longitude": {
                                "type": "number",
                                "minimum": -180,
                                "maximum": 180
                              }
                            },
                            "required": [
                              "latitude",
                              "longitude"
                            ],
                            "additionalProperties": false,
                            "nullable": true
                          },
                          "mapsUrl": {
                            "type": "string",
                            "nullable": true
                          },
                          "openNow": {
                            "type": "boolean",
                            "nullable": true
                          },
                          "priceLevel": {
                            "type": "string",
                            "nullable": true
                          },
                          "rating": {
                            "type": "number"
                          },
                          "types": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "businessStatus": {
                            "type": "string"
                          },
                          "utcOffsetMinutes": {
                            "type": "number"
                          },
                          "priceRange": {
                            "type": "object",
                            "properties": {
                              "low": {
                                "type": "number",
                                "minimum": 0
                              },
                              "high": {
                                "type": "number",
                                "minimum": 0
                              },
                              "currency": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "low",
                              "high",
                              "currency"
                            ],
                            "additionalProperties": false
                          },
                          "openingPeriods": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "open": {
                                  "type": "object",
                                  "properties": {
                                    "day": {
                                      "type": "number"
                                    },
                                    "hour": {
                                      "type": "number"
                                    },
                                    "minute": {
                                      "type": "number"
                                    },
                                    "date": {
                                      "type": "object",
                                      "properties": {
                                        "year": {
                                          "type": "number"
                                        },
                                        "month": {
                                          "type": "number"
                                        },
                                        "day": {
                                          "type": "number"
                                        }
                                      },
                                      "required": [
                                        "year",
                                        "month",
                                        "day"
                                      ],
                                      "additionalProperties": false
                                    }
                                  },
                                  "additionalProperties": false
                                },
                                "close": {
                                  "type": "object",
                                  "properties": {
                                    "day": {
                                      "type": "number"
                                    },
                                    "hour": {
                                      "type": "number"
                                    },
                                    "minute": {
                                      "type": "number"
                                    },
                                    "date": {
                                      "type": "object",
                                      "properties": {
                                        "year": {
                                          "type": "number"
                                        },
                                        "month": {
                                          "type": "number"
                                        },
                                        "day": {
                                          "type": "number"
                                        }
                                      },
                                      "required": [
                                        "year",
                                        "month",
                                        "day"
                                      ],
                                      "additionalProperties": false
                                    }
                                  },
                                  "additionalProperties": false
                                }
                              },
                              "required": [
                                "open"
                              ],
                              "additionalProperties": false
                            }
                          },
                          "attributions": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "provider": {
                                  "type": "string"
                                },
                                "providerUri": {
                                  "type": "string",
                                  "nullable": true
                                }
                              },
                              "required": [
                                "provider",
                                "providerUri"
                              ],
                              "additionalProperties": false
                            }
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "address",
                          "location",
                          "mapsUrl",
                          "openNow",
                          "priceLevel",
                          "attributions"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "warnings": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "dataMode",
                    "places",
                    "warnings"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/places/search": {
      "post": {
        "summary": "Resolve typed destinations with a Miami location bias",
        "tags": [
          "Places"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "query": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 200
                  }
                },
                "required": [
                  "query"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "security": [],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "dataMode": {
                      "type": "string",
                      "enum": [
                        "demo",
                        "live"
                      ]
                    },
                    "places": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "address": {
                            "type": "string"
                          },
                          "location": {
                            "type": "object",
                            "properties": {
                              "latitude": {
                                "type": "number",
                                "minimum": -90,
                                "maximum": 90
                              },
                              "longitude": {
                                "type": "number",
                                "minimum": -180,
                                "maximum": 180
                              }
                            },
                            "required": [
                              "latitude",
                              "longitude"
                            ],
                            "additionalProperties": false,
                            "nullable": true
                          },
                          "mapsUrl": {
                            "type": "string",
                            "nullable": true
                          },
                          "openNow": {
                            "type": "boolean",
                            "nullable": true
                          },
                          "priceLevel": {
                            "type": "string",
                            "nullable": true
                          },
                          "rating": {
                            "type": "number"
                          },
                          "types": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "businessStatus": {
                            "type": "string"
                          },
                          "utcOffsetMinutes": {
                            "type": "number"
                          },
                          "priceRange": {
                            "type": "object",
                            "properties": {
                              "low": {
                                "type": "number",
                                "minimum": 0
                              },
                              "high": {
                                "type": "number",
                                "minimum": 0
                              },
                              "currency": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "low",
                              "high",
                              "currency"
                            ],
                            "additionalProperties": false
                          },
                          "openingPeriods": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "open": {
                                  "type": "object",
                                  "properties": {
                                    "day": {
                                      "type": "number"
                                    },
                                    "hour": {
                                      "type": "number"
                                    },
                                    "minute": {
                                      "type": "number"
                                    },
                                    "date": {
                                      "type": "object",
                                      "properties": {
                                        "year": {
                                          "type": "number"
                                        },
                                        "month": {
                                          "type": "number"
                                        },
                                        "day": {
                                          "type": "number"
                                        }
                                      },
                                      "required": [
                                        "year",
                                        "month",
                                        "day"
                                      ],
                                      "additionalProperties": false
                                    }
                                  },
                                  "additionalProperties": false
                                },
                                "close": {
                                  "type": "object",
                                  "properties": {
                                    "day": {
                                      "type": "number"
                                    },
                                    "hour": {
                                      "type": "number"
                                    },
                                    "minute": {
                                      "type": "number"
                                    },
                                    "date": {
                                      "type": "object",
                                      "properties": {
                                        "year": {
                                          "type": "number"
                                        },
                                        "month": {
                                          "type": "number"
                                        },
                                        "day": {
                                          "type": "number"
                                        }
                                      },
                                      "required": [
                                        "year",
                                        "month",
                                        "day"
                                      ],
                                      "additionalProperties": false
                                    }
                                  },
                                  "additionalProperties": false
                                }
                              },
                              "required": [
                                "open"
                              ],
                              "additionalProperties": false
                            }
                          },
                          "attributions": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "provider": {
                                  "type": "string"
                                },
                                "providerUri": {
                                  "type": "string",
                                  "nullable": true
                                }
                              },
                              "required": [
                                "provider",
                                "providerUri"
                              ],
                              "additionalProperties": false
                            }
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "address",
                          "location",
                          "mapsUrl",
                          "openNow",
                          "priceLevel",
                          "attributions"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "warnings": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "dataMode",
                    "places",
                    "warnings"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    }
  }
}